requirements.txt

MetaTrader5
pandas
numpy
openai
fastapi
uvicorn
python-dotenv
websockets
